# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Wedding",
        "description": "Pra casar API for wedding name Pra Casar",
        "termsOfService": "",
        "contact": {
            "email": "admin@approx.com.br"
        },
        "license": {
            "name": "Pro Produtores",
            "url": "https://www.proprodutores.com.br/"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development server"
        }
    ],
    "paths": {
        "/action": {
            "post": {
                "tags": [
                    "Action"
                ],
                "summary": "Create Chroma Image",
                "description": "Create Chroma image",
                "operationId": "App\\Http\\Controllers\\ActionController::create",
                "requestBody": {
                    "description": "Adding Chroma Image to Database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChromaImage"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/actions": {
            "get": {
                "tags": [
                    "Action"
                ],
                "summary": "Get list of Actions",
                "description": "Returns list of Actions",
                "operationId": "App\\Http\\Controllers\\ActionController::index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Action"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/chroma/image": {
            "get": {
                "tags": [
                    "ChromaImage"
                ],
                "summary": "List Chroma Image",
                "description": "Return the List of Chroma image per User",
                "operationId": "App\\Http\\Controllers\\ChromaImageController::index",
                "parameters": "",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChromaImage"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ChromaImage"
                ],
                "summary": "Create Chroma Image",
                "description": "Create Chroma image",
                "operationId": "App\\Http\\Controllers\\ChromaImageController::create",
                "requestBody": {
                    "description": "Adding Chroma Image to Database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChromaImage"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/chroma/image{image}": {
            "delete": {
                "tags": [
                    "ChromaImage"
                ],
                "summary": "Delete Chroma Image",
                "description": "Delete Chroma image of the User",
                "operationId": "App\\Http\\Controllers\\ChromaImageController::delete",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "Chroma Image of the User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/post/{post}/like": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Post Likes",
                "description": "Save the Post Likes on pivot table by post id",
                "operationId": "App\\Http\\Controllers\\PostController::postLike",
                "requestBody": {
                    "description": "Post attributes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostLike"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Action": {
                "required": [
                    "id",
                    "title",
                    "expense",
                    "notify_guests",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Cha de Panela"
                    },
                    "expense": {
                        "type": "tinyint",
                        "example": 1
                    },
                    "expense_value": {
                        "type": "double",
                        "example": 850.06
                    },
                    "expense_date": {
                        "type": "date",
                        "example": "17/01/2018"
                    },
                    "notify_guests": {
                        "type": "tinyint",
                        "example": 1
                    },
                    "notify_date_from": {
                        "type": "date",
                        "example": "17/01/2018"
                    },
                    "notify_date_to": {
                        "type": "date",
                        "example": "17/02/2018"
                    },
                    "message": {
                        "type": "string",
                        "example": "Compareça"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "ChromaImage": {
                "required": [
                    "img_url",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "img_url": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "PostLike": {
                "required": [
                    "post_id",
                    "guest_id"
                ],
                "properties": {
                    "post_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "guest_id": {
                        "type": "string",
                        "example": "JMF87676"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}